fix(item-transfer): 完善物品反查与识别容错
- 为库存转移动作补充画面稳定等待，并处理物品反查的结构化错误
- 将网格检测未命中改为可恢复结果，避免画面识别失败通过异常传播
- 拆分物品识别 Go 公共类型，统一参数、过滤器和返回错误定义
