+++ assets/resource/pipeline/ItemTransfer.json
@@ -243,7 +243,10 @@
         ],
         "expected": "ORIGIN_REGION",
         "action": "Click",
-        "post_delay": 400,
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000
+        },
         "next": [
             "ItemTransferClickOriginClicked"
         ]
@@ -746,6 +749,16 @@
             -52,
             -50
         ],
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000,
+            "target": [
+                154,
+                202,
+                585,
+                291
+            ]
+        },
         "next": [
             "[Anchor]ItemTransferAfterRepoTransfer"
         ]
@@ -760,6 +773,16 @@
             -52,
             -50
         ],
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000,
+            "target": [
+                154,
+                202,
+                585,
+                291
+            ]
+        },
         "next": [
             "[Anchor]ItemTransferAfterRepoTransfer"
         ]
@@ -790,7 +813,10 @@
         ],
         "expected": "DESTINATION_REGION",
         "action": "Click",
-        "post_delay": 400,
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000
+        },
         "next": [
             "ItemTransferClickDestinationClicked"
         ]
@@ -1078,6 +1104,16 @@
             -52,
             -50
         ],
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000,
+            "target": [
+                739,
+                202,
+                398,
+                291
+            ]
+        },
         "next": [
             "[Anchor]ItemTransferAfterBagTransfer"
         ]
@@ -1092,6 +1128,16 @@
             -52,
             -50
         ],
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000,
+            "target": [
+                739,
+                202,
+                398,
+                291
+            ]
+        },
         "next": [
             "[Anchor]ItemTransferAfterBagTransfer"
         ]
@@ -1336,7 +1382,10 @@
         ],
         "expected": "ORIGIN_REGION",
         "action": "Click",
-        "post_delay": 400,
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000
+        },
         "next": [
             "ItemTransferClickOriginReturnClicked"
         ]
@@ -1492,6 +1541,16 @@
             -52,
             -50
         ],
+        "post_wait_freezes": {
+            "time": 400,
+            "timeout": 3000,
+            "target": [
+                739,
+                202,
+                398,
+                291
+            ]
+        },
         "next": [
             "ItemTransferCheckIfFullReturn",
             "ItemTransferBagReturnCompleted"
